Add startup checking to nodejs daemon
The daemon start portion of the deploy recipe wasn't checking if the daemon was already running and consequently failing when it was. This adds a check which installs/starts the daemon if not running, otherwise restarts the running daemon.

I also renamed the shutdown recipe to undeploy to match the phase in opswork and added a check for the daemon to be running before trying to stop it.